Simplify the main function.
