+++ CMakeLists.txt
@@ -3,4 +3,4 @@ project(AST)
 
 set(CMAKE_CXX_STANDARD 14)
 
-add_executable(AST main.cpp)
+add_executable(AST main.cpp Pretreatment.cpp Pretreatment.h)
